rename: 경매 상세 페이지 폴더 및 파일명 변경
