+++ apps/web/app/auction/[id]/page.tsx
@@ -1,0 +1,7 @@
+import AuctionPage from '@/pages/AuctionPage';
+
+const Page = () => {
+  return <AuctionPage />;
+};
+
+export default Page;
