--- apps/web/app/product/[id]/page.tsx
@@ -1,7 +1,0 @@
-import ProductDetailPage from '@/pages/ProductDetailPage';
-
-const Page = () => {
-  return <ProductDetailPage />;
-};
-
-export default Page;
