+++ apps/web/src/components/auction-detail/banner-direct.tsx
