+++ apps/web/src/components/auction-detail/bid-form.tsx
@@ -1,9 +1,20 @@
 import { Button } from '@/components/ui/button';
-import ButtonFavorite from '@/components/item-detail/button-favorite';
-import BannerDirect from '@/components/item-detail/banner-direct';
+import ButtonFavorite from '@/components/auction-detail/button-favorite';
+import BannerDirect from '@/components/auction-detail/banner-direct';
 import { formatCurrencyWithUnit } from '@/utils/price';
-import BidInput from '@/components/item-detail/bid-input';
+import BidInput from '@/components/auction-detail/bid-input';
 import { useState } from 'react';
+import { tv } from 'tailwind-variants';
+
+const bidInputWrapper = tv({
+  base: 'transition-all duration-700 overflow-hidden',
+  variants: {
+    open: {
+      true: 'pt-2 max-h-[300px] translate-y-0',
+      false: 'max-h-0 translate-y-4 pointer-events-none',
+    },
+  },
+});
 
 const BidForm = () => {
   const [isBidding, setIsBidding] = useState(false);
@@ -19,11 +30,21 @@ const BidForm = () => {
   };
 
   return (
-    <div className="flex w-full flex-col gap-4 bg-white p-4 shadow-[0_-4px_8px_-2px_rgba(0,0,0,0.05)]">
-      <BannerDirect directPrice={directPrice} />
-      {isBidding && (
-        <BidInput currentPrice={15000} minUnit={1000} bidPrice={bidPrice} onChange={setBidPrice} />
-      )}
+    <div className="duration-600 flex w-full flex-col gap-2 bg-white p-4 shadow-[0_-4px_8px_-2px_rgba(0,0,0,0.05)] transition-all">
+      <div className="flex w-full justify-center">
+        <div className="h-1 w-1/3 rounded-full bg-neutral-200" />
+      </div>
+      <div className={`duration-600 transition-all ${isBidding ? 'pb-4' : 'pb-0'}`}>
+        <BannerDirect directPrice={directPrice} />
+        <div className={bidInputWrapper({ open: isBidding })}>
+          <BidInput
+            currentPrice={15000}
+            minUnit={1000}
+            bidPrice={bidPrice}
+            onChange={setBidPrice}
+          />
+        </div>
+      </div>
       <div className="flex items-center justify-between">
         <div className="flex items-center gap-2">
           <ButtonFavorite itemId="123" />
