+++ apps/web/src/components/auction-detail/bid-input.tsx
