+++ apps/web/src/components/auction-detail/bid-table.tsx
