+++ apps/web/src/components/auction-detail/item-information.tsx
@@ -6,7 +6,7 @@ interface ItemInformationProps {
   item: Item;
 }
 
-interface Item {
+export interface Item {
   title: string;
   status: string;
   endTime: string;
@@ -16,7 +16,7 @@ const ItemInformation = ({ item }: ItemInformationProps) => {
   const deadline = formatDate(new Date(item.endTime));
 
   return (
-    <section className="flex w-full flex-col p-4">
+    <section className="flex h-auto w-full flex-col border-b border-t border-neutral-100 bg-white p-4">
       <Badge className="rounded-sm px-1.5 py-0 text-xs">{item.status}</Badge>
       <h2 className="mt-1.5 text-xl font-semibold text-neutral-900">{item.title}</h2>
       <div className="mt-2 flex justify-between">
